gen_purpose_mqtt: moved and modified gen purpose MQTT
Moved the MQTT implementation I was using for the python postprocess engine to put it at the root of the directory src/ so that other people can use it too. I modified it a bit so that it could be used to implement either a publisher or a subscriber, thus making it general purpose.