File structure and first code for Fboard pp engine
I created the file structure and a first version of the Freeboard post processing machine. The FboardAdapter should be ready like this. Due to technical difficulties testing will start once alex is back from his exam... the tools will need to be modified and so far only work to be able to try the adapter.